Added player, no movement

+++ brain/src/render/axis_state.rs
@@ -8,15 +8,15 @@ pub enum Side {
 }
 
 pub struct AxisState {
-    pub map_pos: i32,
-    pub step: i32,
+    pub map_pos: i64,
+    pub step: i64,
     pub side_dist: f32,
     pub delta_dist: f32,
 }
 
 impl AxisState {
     pub fn new(origin: f32, dir: f32) -> Self {
-        let map_pos = origin.floor() as i32;
+        let map_pos = origin.floor() as i64;
         let delta_dist = (1.0 / dir).abs();
 
         let (step, side_dist) = if dir < 0.0 {
